[minor] Extract common util struct for table event sync
