+++ src/moonlink/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod error;
+pub mod event_sync;
 pub mod row;
 mod storage;
 mod table_handler;
@@ -6,6 +7,7 @@ pub(crate) mod table_notify;
 mod union_read;
 
 pub use error::*;
+pub use event_sync::EventSyncSender;
 pub use storage::storage_utils::create_data_file;
 pub(crate) use storage::NonEvictableHandle;
 pub use storage::SnapshotReadOutput;
@@ -17,7 +19,7 @@ pub use storage::{
     FileSystemAccessor, MoonlinkSecretType, MoonlinkTableConfig, MoonlinkTableSecret,
     ObjectStorageCache, ObjectStorageCacheConfig,
 };
-pub use table_handler::{EventSyncSender, TableHandler};
+pub use table_handler::TableHandler;
 pub use table_notify::TableEvent;
 pub use union_read::{ReadState, ReadStateManager};
